PI DISP - Add generic functions for display
