+++ pi-display-foundation/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2024"
 
 [dependencies]
 embedded-graphics = "0.8.0"
+tinybmp = "0.6.0"
 midir = "0.10.2"
 crossbeam-channel = "0.5.15"
 triple_buffer = "8.1.1"
@@ -12,4 +13,3 @@ triple_buffer = "8.1.1"
 [target.'cfg(target_os = "linux")'.dependencies]
 sh1106 = "0.5.0"
 linux-embedded-hal = "0.3.0"
-tinybmp = "0.6.0"
