+++ pi-display-foundation/src/modules.rs
@@ -3,3 +3,4 @@ pub mod state;
 pub mod midi;
 pub mod coordinator;
 pub mod display;
+pub mod screens;
